deploy: 메인 배포 테스트
