+++ app/products/page.tsx
@@ -10,7 +10,7 @@ export const metadata: Metadata = {
     description: '카테고리별 상품을 한눈에 확인하세요.',
     images: [
       {
-        url: '/og/products.png',
+        url: '/og/global.png',
         width: 1200,
         height: 630,
       },
